+++ src/client/cl_demo.cpp
@@ -498,7 +498,7 @@ void __cdecl CL_Record_f()
                     v4 = SL_ConvertToString(*configstrings);
                     MSG_WriteString(&v13, (char*)v4);
                     ++configstrings;
-                } while ((int)configstrings < (int)clients[0].mapname);
+                } while ((uintptr_t)configstrings < (uintptr_t)clients[0].mapname);
                 v5 = SL_ConvertToString(clients[0].configstrings[0]);
                 v6 = (char*)Info_ValueForKey(v5, "mapname");
                 v7 = v6;
