+++ src/client/cl_parse.cpp
@@ -370,7 +370,7 @@ void __cdecl CL_ParseGamestate(char *configstrings)
     LABEL_15:
         ++v3;
         ++v2;
-    } while ((int)v3 < (int)clients[0].mapname);
+    } while ((uintptr_t)v3 < (uintptr_t)clients[0].mapname);
 }
 
 void __cdecl CL_ParseServerCommands(msg_t *msg)
