+++ src/game/g_save.cpp
@@ -1611,7 +1611,7 @@ void __cdecl G_LoadModelPrecacheList(SaveGame *save)
         else
             v4 = 0;
         *modelMap++ = v4;
-    } while ((int)modelMap < (int)&level.priorityNodeBias);
+    } while ((uintptr_t)modelMap < (uintptr_t)&level.priorityNodeBias);
 }
 
 void __cdecl G_ClearConfigstrings(int iFirst, int iCount)
@@ -1744,7 +1744,7 @@ void __cdecl G_SaveWeaponCue(SaveGame *save)
             v4 = 0;
         SaveMemory_SaveWrite(&v4, 4, save);
         ++droppedWeaponCue;
-    } while ((int)droppedWeaponCue < (int)&level.changelevel);
+    } while ((uintptr_t)droppedWeaponCue < (uintptr_t)&level.changelevel);
 }
 
 void __cdecl G_LoadWeaponCue(SaveGame *save)
@@ -1787,7 +1787,7 @@ void __cdecl G_LoadWeaponCue(SaveGame *save)
         if (!v5)
             droppedWeaponCue->setEnt(&g_entities[v4 - 1]);
         ++droppedWeaponCue;
-    } while ((int)droppedWeaponCue < (int)&level.changelevel);
+    } while ((uintptr_t)droppedWeaponCue < (uintptr_t)&level.changelevel);
 }
 
 void __cdecl G_SaveDvars(SaveGame *save)
